refactor(app): extract community bridge
